Crate rustorio

Modules

buildings

Buildings take inputs to produce outputs over time.

gamemodes

recipes

A recipe is a way of turning resources into other resources. A specific recipe specifies the input and output resources, as well as the time it takes to complete the recipe.

Structs

Bundle

Contains a fixed (compile-time known) amount of a resource. A Bundle can be used to build structures or as input for recipes.

Resource

Holds an arbitrary amount of a resource. A Resource object can be split into smaller parts, combined or Bundles can be extracted from them.

Tick

The tick is used to keep track of time in the game. You can advance the game by one tick using the next method. Many functions and building methods require a Tick to be passed in, which allows them to update their state. If a function takes a &mut Tick, then the function will take time. If a function merely takes a &Tick, or no Tick at all, it will never advance the game time.

Enums

ResourceType

Functions

mine_copper

Mines copper ore. Takes 2 ticks to mine 1 ore.

mine_iron

Mines iron ore. Takes 2 ticks to mine 1 ore.

play

Runs your play. If it is run multiple times, it will panic. This is to prevent using multiple threads to cheat.

Type Aliases

CopperOreBundle

IronOreBundle