rustorio

Module buildings

Buildings take inputs to produce outputs over time.

To use a building, you must first build it which takes a number of resources. Then you can add inputs to it using add_input or similar functions. Once it has sufficient inputs, it will start producing outputs, which can be extracted using take_output or similar functions.

When created, a building is set to a specific Recipe, which defines the inputs and outputs. This can be changed using the change_recipe method, but only if the building is empty (no inputs or outputs).

Structs

Assembler

The assembler is used for recipes that require two different inputs to produce an output.

Furnace

The furnace is used to smelt ores into base resources.