rustorio

Function play

pub fn play<G: GameMode>(
    main: fn(Tick, G::StartingResources) -> (Tick, G::VictoryResources),
) -> !

Runs your play. If it is run multiple times, it will panic. This is to prevent using multiple threads to cheat.